test-requirements.txt
cloudify-common==6.4.0

mock>=1.0
nose>=1.3
coverage
testtools>=0.9.36,!=1.2.0
tox
nose-cov
pylint
flake8
requests-mock
PyYAML
xmltodict
isort